+++ packages/networks/src/networks/goerli.ts
@@ -108,6 +108,11 @@ export const goerli: NetworkConfig = {
         name: 'Guild',
         address: '0x1AC9271D271b8E50537CAd54b330424C52A84822',
       },
+      {
+        id: HookType.PROMOCODE,
+        name: 'Discount code',
+        address: '0x850c015A6A88756a59Dc025fca988494fF90DBB7',
+      },
     ],
   },
 }